requirements.txt
flask
gunicorn
werkzeug
numpy
scipy<1.9
librosa<0.10.0
soundfile
pydub
mido
midi2audio
torch
torchaudio
basic-pitch
torchcrepe
piano-transcription-inference
demucs
requests
roex-python
tensorflow
onnxruntime
tenacity
git+https://github.com/algorithmic-music-exploration/amen.git


musicai-sdk
music21
textblob
google-auth
google-auth-oauthlib
google-generativeai
colorama
midiutil
PyYAML
ruamel.yaml
